moved all plotting functions to separate TweezerPlotter and BulkGasPlotter classes in separate files for modularity
